Move EmptyReport and UnknownMessage to pylint.exceptions
