+++ doc/whatsnew/2.0.rst
@@ -145,6 +145,8 @@ Other Changes
 * Add ``InvalidMessageError`` exception class and replace ``assert`` in
   pylint.utils with ``raise InvalidMessageError``.
 
+* ``UnknownMessage`` and ``EmptyReport`` are now provided by the new
+  ``pylint.exceptions`` submodule instead of ``pylint.utils`` as before.
 
 Bug fixes
 =========
