+++ pylint/exceptions.py
@@ -6,3 +6,9 @@
 
 class InvalidMessageError(Exception):
     """raised when a message creation, registration or addition is rejected"""
+
+class UnknownMessage(Exception):
+    """raised when a unregistered message id is encountered"""
+
+class EmptyReport(Exception):
+    """raised when a report is empty and so should not be displayed"""
